fix: automatically install vcredist if missing, move uv to python directory, automatically set --cpu if no other backend is available

+++ packages/metastable/src/setup/tasks/extract.ts
@@ -9,15 +9,18 @@ export class ExtractTask extends BaseTask {
   constructor(
     private archivePath: string,
     private targetPath: string,
+    private cleanDir = false,
   ) {
     super('extract', undefined);
   }
 
   async execute() {
-    this.appendLog('Cleaning up...');
-    try {
-      await rimraf(this.targetPath);
-    } catch {}
+    if (this.cleanDir) {
+      this.appendLog('Cleaning up...');
+      try {
+        await rimraf(this.targetPath);
+      } catch {}
+    }
 
     this.appendLog(`Extracting ${this.archivePath} to ${this.targetPath}`);
